Split Client interface/implementation by files
