+++ src/controller.ts
@@ -1,7 +1,8 @@
 #!/usr/bin/env node
 
 import { supplyClient } from './ZeroconfClient'
-import { Client, Http2Client } from './Client'
+import { Client } from './Client'
+import { Http2Client } from './Http2Client'
 
 import * as bonjour from 'bonjour'
 import { once } from 'events'
